test(federated): consolidate test files into focused modules
- Consolidated federated simulation tests from directory structure to single files
- Separated unit tests from integration tests for better parallel execution
- Removed duplicate test classes and helper functions
- Improved test organization with shared utility functions
